requirements.txt
fastapi
uvicorn
nest-asyncio
transformers
torch
pydantic
requests
Pillow
python-multipart